requirements.txt
# Core dependencies
Pillow>=9.0.0
gpiozero>=1.6.2
python-weather>=0.4.2
caldav>=0.9.0
icalevents>=0.1.25
icalendar>=4.1.0
posix_ipc>=1.0.5
pytz>=2021.3
tzlocal>=4.0
humanize>=3.13.1

# Web interface dependencies
Flask>=2.0.0

# Feed reader dependencies (integrated in web/)
feedparser>=6.0.10

# System monitoring
psutil>=5.9.0
distro>=1.7.0

# Webview screen
htmlwebshot>=0.1.0
